+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Split the dual-functionality of the Junk Frame button into two distinct buttons: `Start Selling` & `Destroy Next Item`
+- Added tooltips to the Junk Frame's `Destroy Next Item` button: mousing over the button will display the name of the next item to be destroyed, and holding shift will display the item's tooltip
+
+### Fixed
+
+- Issue with frame levels
+- Support for reagent bag in retail ([#165](https://github.com/moody/Dejunk/issues/165))
+
 ## [1.5.0] - 2023-01-27
 
 ### Added